Move apply code from kubectl to apimachinery
