+++ staging/src/k8s.io/apimachinery/pkg/apply/parse/item.go
@@ -17,8 +17,8 @@ limitations under the License.
 package parse
 
 import (
+	"k8s.io/apimachinery/pkg/apply"
 	"k8s.io/kube-openapi/pkg/util/proto"
-	"k8s.io/kubernetes/pkg/kubectl/apply"
 )
 
 // Item wraps values from 3 sources (recorded, local, remote).
