+++ staging/src/k8s.io/apimachinery/pkg/apply/parse/util.go
@@ -21,8 +21,8 @@ import (
 	"reflect"
 	"strings"
 
+	"k8s.io/apimachinery/pkg/apply"
 	"k8s.io/kube-openapi/pkg/util/proto"
-	"k8s.io/kubernetes/pkg/kubectl/apply"
 )
 
 // nilSafeLookup returns the value from the map if the map is non-nil
