+++ staging/src/k8s.io/apimachinery/pkg/apply/strategy/replace_map_test.go
@@ -19,8 +19,8 @@ package strategy_test
 import (
 	. "github.com/onsi/ginkgo"
 
+	"k8s.io/apimachinery/pkg/apply/strategy"
 	"k8s.io/kube-openapi/pkg/util/proto"
-	"k8s.io/kubernetes/pkg/kubectl/apply/strategy"
 )
 
 var _ = Describe("Replacing fields of type map with openapi for some fields", func() {
