+++ staging/src/k8s.io/apimachinery/pkg/apply/strategy/strategic_visitor.go
@@ -17,7 +17,7 @@ limitations under the License.
 package strategy
 
 import (
-	"k8s.io/kubernetes/pkg/kubectl/apply"
+	"k8s.io/apimachinery/pkg/apply"
 )
 
 // delegatingStrategy delegates merging fields to other visitor implementations
